requirements.txt
asgiref==3.8.1
billiard==4.2.1
build==1.2.2.post1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
flower==2.0.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pipdeptree==2.24.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
